+++ apps/editor-dev/pages/plugins/CenterPlugin.jsx
@@ -1,20 +1,24 @@
 import { View } from "@elf-framework/ui";
 
-export async function CenterPlugin(editor) {
-  const { Canvas3D } = await import("./CanvasView");
+import { Canvas3D } from "./CanvasView";
 
+export function RenderView() {
+  return (
+    <View position="relative" height="100%" overflow="hidden">
+      <Canvas3D />
+      <button
+        type="button"
+        style={{ position: "absolute" }}
+        onClick={() => console.log("Hello World")}
+      >
+        Hello World
+      </button>
+    </View>
+  );
+}
+
+export async function CenterPlugin(editor) {
   editor.registerGroupUI("center-panel", {
-    renderView: (
-      <View position="relative" height="100%" overflow="hidden">
-        <Canvas3D />
-        <button
-          type="button"
-          style={{ position: "absolute" }}
-          onClick={() => console.log(editor.configs.get("yellow"))}
-        >
-          Hello
-        </button>
-      </View>
-    ),
+    renderView: <RenderView />,
   });
 }