+++ packages/base-editor/src/components/status/Loading.jsx
@@ -1,0 +1,18 @@
+export function Loading() {
+  return (
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        height: "100%",
+        width: "100%",
+        position: "absolute",
+        top: 0,
+        left: 0,
+      }}
+    >
+      Loading...
+    </div>
+  );
+}