+++ apps/editor-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # editor-dev
 
+## 1.0.9
+
+### Patch Changes
+
+- modify plugin structure for hmr
+
 ## 1.0.8
 
 ### Patch Changes